Добавляет вспомгательную функцию, корректирует название файлов
