+++ src/utils.js
@@ -1,10 +1,6 @@
-export const renderPosition = {
-  BEFOREBEGIN: 'beforebegin',
-  AFTERBEGIN: 'afterbegin',
-  BEFOREEND: 'beforeend',
-  AFTEREND: 'afterend'
-};
+export const getRandomInteger = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
 
-export const renderTemplate = (container, template, place) => {
-  container.insertAdjacentHTML(place, template);
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
 };
